Small cleanups in testbed, hide unnecessary public structs
