+++ testbed/tests/e2e_test.go
@@ -35,7 +35,7 @@ func TestIdleMode(t *testing.T) {
 		dataProvider,
 		testbed.NewJaegerGRPCDataSender(testbed.DefaultHost, testbed.DefaultJaegerPort),
 		testbed.NewOCDataReceiver(testbed.DefaultOCPort),
-		&testbed.ChildProcess{},
+		testbed.NewChildProcessCollector(),
 		&testbed.PerfTestValidator{},
 		performanceResultsSummary,
 		testbed.WithResourceLimits(testbed.ResourceSpec{ExpectedMaxCPU: 4, ExpectedMaxRAM: 70}),
@@ -65,7 +65,7 @@ func TestBallastMemory(t *testing.T) {
 			dataProvider,
 			testbed.NewJaegerGRPCDataSender(testbed.DefaultHost, testbed.DefaultJaegerPort),
 			testbed.NewOCDataReceiver(testbed.DefaultOCPort),
-			&testbed.ChildProcess{},
+			testbed.NewChildProcessCollector(),
 			&testbed.PerfTestValidator{},
 			performanceResultsSummary,
 			testbed.WithSkipResults(),
